pyproject.toml
[tool.poetry]
name = 'air_bot'
version = '0.1.0'
description = 'Simple chat bot created to bridge information flow between the Air project and the Wh00t chat app'
authors = ['roboto84 <7658704+roboto84@users.noreply.github.com>']
license = 'GPL-3.0-or-later'

[tool.poetry.dependencies]
python = '^3.8'
python-dotenv = '^0.15.0'
APScheduler = '^3.6.3'
air-core = {git = 'https://github.com/roboto84/air_core.git', rev = '432f76fe'}
wh00t-core = {git = 'https://github.com/roboto84/wh00t_core.git', tag = 'v0.4.0'}

[tool.poetry.dev-dependencies]
pytest = '^5.2'

[build-system]
requires = ['poetry-core>=1.0.0']
build-backend = 'poetry.core.masonry.api'
